RAGUdemy/requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
faiss-cpu
tiktoken
chromadb
pypdf
pymupdf
unstructured
python-docx
docx2txt
python-dotenv
pandas
ipykernel
ipywidgets
openpyxl
msoffcrypto-tool
jq
matplotlib
langchain-huggingface
sentence-transformers
langchain-astradb
langchain-pinecone
langchain-experimental
rank_bm25
langchain-google-genai
langchain-tavily
arxiv
wikipedia
langgraph
langgraph-cli[inmem]